+++ lib/gooddata/lcm/actions/purge_clients.rb
@@ -12,14 +12,6 @@ module GoodData
       DESCRIPTION = 'Purge LCM Clients'
 
       PARAMS = define_params(self) do
-        description 'Client Used for Connecting to GD'
-        param :gdc_gd_client, instance_of(Type::GdClientType), required: true
-
-        description 'Organization Name'
-        param :organization, instance_of(Type::StringType), required: true
-
-        description 'Segments to manage'
-        param :segments, array_of(instance_of(Type::SegmentType)), required: true
       end
 
       RESULT_HEADER = [
@@ -30,41 +22,35 @@ module GoodData
 
       class << self
         def call(params)
-          client = params.gdc_gd_client
+          client_projects = params.client_projects
 
-          domain_name = params.organization || params.domain
-          domain = client.domain(domain_name) || fail("Invalid domain name specified - #{domain_name}")
-          all_segments = domain.segments
+          results = client_projects.pmap do |_, info|
+            client = info[:segment_client]
+            project = info[:project]
 
-          segment_names = params.segments.map do |segment|
-            segment.segment_id.downcase
-          end
+            res = {
+              client_id: client.client_id,
+              project: project && project.pid
+            }
 
-          segments = all_segments.select do |segment|
-            segment_names.include?(segment.segment_id.downcase)
-          end
-
-          results = segments.pmap do |segment|
-            segment.clients.map do |segment_client|
-              project = segment_client.project
-              res = {
-                client_id: segment_client.client_id,
-                project: project && project.pid
-              }
-
-              if project.nil? || project.deleted?
-                segment_client.delete
-                res[:status] = 'purged'
-              else
-                res[:status] = 'ok - not purged'
-              end
-
-              res
+            if project.nil? || project.deleted?
+              client.delete
+              res[:status] = 'purged'
+            else
+              res[:status] = 'ok - not purged'
             end
+
+            res
           end
 
-          # Return results
-          results.flatten
+          results.pselect { |res| res[:status] == 'purged' }.pmap { |res| res[:client_id] }.each { |id| client_projects.delete(id) }
+
+          {
+            results: results,
+            params: {
+              client_projects: client_projects
+            }
+          }
         end
       end
     end