+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - TMA-563 - Provisioning brick doesn't verify status of project_id
 - TMA-502 - User and User Filters brick: multiple_projects_column = client_id should be default for client modes
 - TMA-565 - Provisioning brick doesn't add technical users when input source contains project_id
+- TMA-547 - Provisioning brick: sync client workspace title
 
 ## 0.6.53
 - TMA-522 - Rollout: Incorrect CLIENT_ID assigned to client schedule